apps/auth-server/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0  # For database migrations
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Authentication
bcrypt>=4.1.2
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0  # Alternative JWT library with more features

# Environment and configuration
python-dotenv>=1.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development tools (optional)
black>=24.0.0
isort>=5.13.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # For testing FastAPI endpoints
